+++ gens/blueprints/__init__.py
@@ -1,4 +1,3 @@
 from .gens.views import gens_bp
 from .home.views import home_bp
 from .login.views import login_bp
-