+++ gens/db/db.py
@@ -1,4 +1,5 @@
 """Functions for handeling database connection."""
+
 import logging
 import os
 
@@ -7,6 +8,8 @@ from pymongo import MongoClient
 
 from gens.exceptions import ConfigurationException
 
+from ..config import settings
+
 LOG = logging.getLogger(__name__)
 
 
@@ -14,16 +17,11 @@ def init_database_connection() -> None:
     """Initialize database connection and store variables to the two databases."""
     # verify that database was properly configured
     LOG.info("Initialize db connection")
-    variables = {}
-    for var_name in ["MONGODB_SCOUT_URI", "MONGODB_GENS_URI", "SCOUT_DBNAME", "GENS_DBNAME"]:
-        if not any([var_name in os.environ, var_name in app.config]):
-            raise ConfigurationException(
-                f"Variable {var_name} not defined in either config or env variable"
-            )
-        variables[var_name] = os.environ.get(var_name, app.config.get(var_name))
+
     # connect to database
-    scout_client = MongoClient(variables["MONGODB_SCOUT_URI"])
-    gens_client = MongoClient(variables["MONGODB_GENS_URI"])
-    # store db handlers in configuration
-    app.config["SCOUT_DB"] = scout_client[variables["SCOUT_DBNAME"]]
-    app.config["GENS_DB"] = gens_client[variables["GENS_DBNAME"]]
+    app.config["SCOUT_DB"] = MongoClient(str(settings.scout_db)).get_database(
+        name=settings.scout_dbname
+    )
+    app.config["GENS_DB"] = MongoClient(str(settings.gens_db)).get_database(
+        name=settings.gens_dbname
+    )