+++ gens/db/samples.py
@@ -1,15 +1,13 @@
 """Store and retrive samples from the database."""
 
-import datetime
 import itertools
 import logging
-from typing import Optional
 
-from pymongo import MongoClient
-from pymongo import DESCENDING
+from pymongo import DESCENDING, MongoClient
 from pymongo.errors import DuplicateKeyError
 
-from .models import SampleObj
+from gens.models.genomic import GenomeBuild
+from gens.models.sample import SampleInfo
 
 LOG = logging.getLogger(__name__)
 
@@ -36,7 +34,7 @@ def store_sample(
     db: MongoClient,
     sample_id: str,
     case_id: str,
-    genome_build: int,
+    genome_build: GenomeBuild,
     baf: str,
     coverage: str,
     overview: str,
@@ -44,6 +42,15 @@ def store_sample(
 ):
     """Store a new sample in the database."""
     LOG.info(f'Store sample "{sample_id}" in database')
+    sample_obj = SampleInfo(
+        sample_id=sample_id,
+        case_id=case_id,
+        genome_build=genome_build,
+        baf_file=baf,
+        coverage_file=coverage,
+        overview_file=overview,
+    )
+    index_fields = ["baf_index", "coverage_index"]
     if force:
         result = db[COLLECTION].update_one(
             {
@@ -51,17 +58,7 @@ def store_sample(
                 "case_id": case_id,
                 "genome_build": genome_build,
             },
-            {
-                "$set": {
-                    "sample_id": sample_id,
-                    "case_id": case_id,
-                    "baf_file": baf,
-                    "coverage_file": coverage,
-                    "overview_file": overview,
-                    "genome_build": genome_build,
-                    "created_at": datetime.datetime.now(),
-                }
-            },
+            {"$set": sample_obj.model_dump(exclude=index_fields)},
             upsert=True,
         )
         if result.modified_count == 1:
@@ -77,34 +74,26 @@ def store_sample(
             )
     else:
         try:
-            db[COLLECTION].insert_one(
-                {
-                    "sample_id": sample_id,
-                    "case_id": case_id,
-                    "baf_file": baf,
-                    "coverage_file": coverage,
-                    "overview_file": overview,
-                    "genome_build": genome_build,
-                    "created_at": datetime.datetime.now(),
-                }
-            )
+            db[COLLECTION].insert_one(sample_obj.model_dump(exclude=index_fields))
         except DuplicateKeyError:
             LOG.error(
                 f'DuplicateKeyError while storing sample with sample_id="{sample_id}" and case_id="{case_id}" in database.'
             )
 
 
-def get_samples(db: MongoClient, start:int = 0, n_samples: int|None = None) -> tuple[list[SampleObj], int]:
+def get_samples(
+    db: MongoClient, start: int = 0, n_samples: int | None = None
+) -> tuple[list[SampleInfo], int]:
     """
     Get samples stored in the databse.
 
     use n_samples to limit the results to x most recent samples
     """
     results = (
-        SampleObj(
+        SampleInfo(
             sample_id=r["sample_id"],
             case_id=r["case_id"],
-            genome_build=r["genome_build"],
+            genome_build=GenomeBuild(int(r["genome_build"])),
             baf_file=r["baf_file"],
             coverage_file=r["coverage_file"],
             overview_file=r["overview_file"],
@@ -118,7 +107,7 @@ def get_samples(db: MongoClient, start:int = 0, n_samples: int|None = None) -> t
     return results, db[COLLECTION].count_documents({})
 
 
-def query_sample(db: MongoClient, sample_id: str, case_id: str|None, _genome_build: int):
+def query_sample(db: MongoClient, sample_id: str, case_id: str | None) -> SampleInfo:
     """Get a sample with id."""
     result = None
     if case_id is None:
@@ -127,11 +116,13 @@ def query_sample(db: MongoClient, sample_id: str, case_id: str|None, _genome_bui
         result = db[COLLECTION].find_one({"sample_id": sample_id, "case_id": case_id})
 
     if result is None:
-        raise SampleNotFoundError(f'No sample with id: "{sample_id}" in database', sample_id)
-    return SampleObj(
+        raise SampleNotFoundError(
+            f'No sample with id: "{sample_id}" in database', sample_id
+        )
+    return SampleInfo(
         sample_id=result["sample_id"],
         case_id=result["case_id"],
-        genome_build=result["genome_build"],
+        genome_build=GenomeBuild(int(result["genome_build"])),
         baf_file=result["baf_file"],
         coverage_file=result["coverage_file"],
         overview_file=result["overview_file"],