+++ gens/wsgi.py
@@ -1,4 +1,5 @@
 """Function for running the app wsgi."""
+
 from .app import create_app
 
 app = create_app()